+++ src/unit_tests/wazuh_modules/agent_upgrade/test_wm_agent_upgrade_manager.c
@@ -98,7 +98,6 @@ void test_wm_agent_upgrade_listen_messages_upgrade_command(void **state)
 
     expect_value(__wrap_wm_agent_upgrade_process_upgrade_command, agent_ids, agents);
     expect_value(__wrap_wm_agent_upgrade_process_upgrade_command, task, upgrade_task);
-    will_return(__wrap_wm_agent_upgrade_process_upgrade_command, 1);
     will_return(__wrap_wm_agent_upgrade_process_upgrade_command, response);
 
     expect_string(__wrap__mtdebug1, tag, "wazuh-modulesd:agent-upgrade");
@@ -111,9 +110,6 @@ void test_wm_agent_upgrade_listen_messages_upgrade_command(void **state)
     expect_value(__wrap_OS_SendSecureTCP, size, strlen(response));
     expect_string(__wrap_OS_SendSecureTCP, msg, response);
     will_return(__wrap_OS_SendSecureTCP, 0);
-
-    expect_memory(__wrap_wm_agent_upgrade_start_upgrades, manager_configs, config, sizeof(config));
-    will_return(__wrap_wm_agent_upgrade_start_upgrades, 1);
 
     wm_agent_upgrade_listen_messages(config);
 }
@@ -170,7 +166,6 @@ void test_wm_agent_upgrade_listen_messages_upgrade_custom_command(void **state)
 
     expect_value(__wrap_wm_agent_upgrade_process_upgrade_custom_command, agent_ids, agents);
     expect_value(__wrap_wm_agent_upgrade_process_upgrade_custom_command, task, upgrade_custom_task);
-    will_return(__wrap_wm_agent_upgrade_process_upgrade_custom_command, 1);
     will_return(__wrap_wm_agent_upgrade_process_upgrade_custom_command, response);
 
     expect_string(__wrap__mtdebug1, tag, "wazuh-modulesd:agent-upgrade");
@@ -183,9 +178,6 @@ void test_wm_agent_upgrade_listen_messages_upgrade_custom_command(void **state)
     expect_value(__wrap_OS_SendSecureTCP, size, strlen(response));
     expect_string(__wrap_OS_SendSecureTCP, msg, response);
     will_return(__wrap_OS_SendSecureTCP, 0);
-
-    expect_memory(__wrap_wm_agent_upgrade_start_upgrades, manager_configs, config, sizeof(config));
-    will_return(__wrap_wm_agent_upgrade_start_upgrades, 1);
 
     wm_agent_upgrade_listen_messages(config);
 }
